requirements.txt
numpy==1.18.5
pandas==1.0.5
Flask_SQLAlchemy==2.4.4
Flask_WTF==0.14.3
scipy==1.5.0
SQLAlchemy_Utils==0.36.8
Werkzeug==1.0.1
simpletransformers==0.48.3
WTForms==2.3.3
requests==2.24.0
SQLAlchemy==1.3.18
Flask_Login==0.5.0
Flask_Assets==2.0
pyngrok==4.1.12
torch==1.6.0
Flask==1.1.2
beautifulsoup4==4.9.1
python-dotenv==0.14.0
joblib==1.0.0